+++ tests/testthat/test-spFromTable.R
@@ -70,8 +70,8 @@ test_that("mismatched attributes and object number is an error", {
 
 
 library(sp)
-library(dplyr)
-library(raster)
+suppressPackageStartupMessages(library(dplyr))
+suppressPackageStartupMessages(library(raster))
 cl <- spbabel::sptable(rasterToContour(raster(volcano)))
 
 cl0 <- cl %>% filter(object_ == -1) #%>% mutate(object_ = as.character(i)) #%>% slice(2:(n()-1))